+++ crates/e2e/tests/e2e/quote_verification.rs
@@ -171,7 +171,7 @@ async fn test_bypass_verification_for_rfq_quotes(web3: Web3) {
         BigDecimal::zero(),
         Default::default(),
         0,
-        u64::MAX,
+        u32::MAX,
     )
     .await
     .unwrap();
